requirements.txt
fastapi 
uvicorn 
pandas
python-multipart 
pillow 
transformers
torch 
torchvision 
scikit-learn